% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordNormUMAP.R
\name{cordNormUMAP}
\alias{cordNormUMAP}
\title{Generate new cds with UMAP coordinates in pData}
\usage{
cordNormUMAP(cds, nComp = NULL)
}
\arguments{
\item{cds}{a cell data set}

\item{nComp}{a vector consisting of the number of PCs to be used}
}
\value{
Returns a CDS with UMAP coordinates in pData
}
\description{
Generate 2D UMAP coordinates based on count normalized expression matrix and append to CDS
}
\examples{

}
