% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordUMAP.R
\name{cordUMAP}
\alias{cordUMAP}
\title{Generate new cds with UMAP coordinates in pData}
\usage{
cordUMAP(cds, nComp = NULL)
}
\arguments{
\item{cds}{a cell data set}

\item{nComp}{a vector consisting of the number of PCs to be used}
}
\value{
Returns a CDS with UMAP coordinates in pData
}
\description{
Generate a 2D UMAP coordinates and append to CDS
}
\examples{

}
