% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupGeneId.R
\name{lookupGeneId}
\alias{lookupGeneId}
\title{Returns the gene ID given a gene short name as input}
\usage{
lookupGeneId(eset, gene_names)
}
\arguments{
\item{eset}{an expression set or cell data set (CDS)}

\item{gene_names}{a list of gene short names}
}
\value{
Returns a vector of gene IDs
}
\description{
Returns the gene ID given a gene short name as input
}
\examples{

}
