% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookupGeneName.R
\name{lookupGeneName}
\alias{lookupGeneName}
\title{Returns the gene short name given a gene ID as input}
\usage{
lookupGeneName(eset, gene_id)
}
\arguments{
\item{eset}{an expression set or cell data set (CDS)}

\item{gene_id}{a list of gene IDs}
}
\value{
Returns a vector of gene short names
}
\description{
Returns the gene short name given a gene ID as input
}
