% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myUMAPPlotMarkers.R
\name{myUMAPPlotMarkers}
\alias{myUMAPPlotMarkers}
\title{Given an exisiting UMAP embedding of scRNAseq data, returns UMAP embedding colored by the relative expression of marker gene of interest}
\usage{
myUMAPPlotMarkers(cds, markers = NULL, logMode = T, shape_by = NULL,
  scaled = FALSE, x_cord = "UMAP1", y_cord = "UMAP2", size = 0.75)
}
\arguments{
\item{cds}{a cell data set}

\item{markers}{a list of gene short names}

\item{logMode}{choose to log10 transform gene expression counts with a pseudocount of 1}

\item{size}{adjust the size of the points used}

\item{x}{the x coordinates that you would like to use}

\item{y}{the y coordinates that you would like to use}
}
\value{
Returns a facet wrapped UMAP embeddings of the same number as the number of marker genes chosen, colored by relative gene expression
}
\description{
Given an exisiting UMAP embedding of scRNAseq data, returns UMAP embedding colored by the relative expression of marker gene of interest
}
\examples{

}
