% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCpicker.R
\name{PCpicker}
\alias{PCpicker}
\title{Visualize PC variance}
\usage{
PCpicker(cds, num_pcs = NULL)
}
\arguments{
\item{cds}{a cell data set}

\item{num_pcs}{the number of PCs you'd like to calculate}
}
\value{
Returns a screeplot with the amount of variance explained by each PC
}
\description{
Visualize PC variance
}
\examples{

}
