% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_violins.R
\name{plot_stacked_violins}
\alias{plot_stacked_violins}
\title{Returns a stacked violin plot of log2 normalized expression subsetted by items in pData. Compatible with ggplot2}
\usage{
plot_stacked_violins(
  cds = NULL,
  genes = c("NULL"),
  group_by = "cluster",
  assay_name = "counts"
)
}
\arguments{
\item{cds}{a cell data set object}

\item{group_by}{a character string indicating the column in pData used for grouping cells}

\item{assay_name}{a character string used to indicate which expression values to use from the Assay slot in the CDS}

\item{gene}{a vector of gene short names}
}
\value{
Returns a stacked violin plot of normalized expression values subsetted by a category in pData
}
\description{
Returns a stacked violin plot of log2 normalized expression subsetted by items in pData. Compatible with ggplot2
}
