% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPatterns.R
\name{plotPatterns}
\alias{plotPatterns}
\title{Returns UMAP visualizations colored by pattern weight as determined by NNLM::nnmf()}
\usage{
plotPatterns(eset, decomp, nPatterns = c(1:5), x_cord = "UMAP1",
  y_cord = "UMAP2", size = 0.5)
}
\arguments{
\item{eset}{an expression set or cell data set (CDS)}

\item{decomp}{output from NNLM::nnmf()}

\item{nPatterns}{a vector of patterns that you'd like to visualize}

\item{x_cord}{the x cordinates to use for visualization ("UMAP1" by default)}

\item{y_cord}{the y cordinates to use for visualization ("UMAP2" by default)}

\item{size}{the size to use for the points}
}
\value{
Returns UMAP visualizations colored by pattern weight as determined by NNLM::nnmf()
}
\description{
Returns UMAP visualizations colored by pattern weight as determined by NNLM::nnmf()
}
\examples{

}
