% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Generate a UMAP visualiztion using ggplot2}
\usage{
plotUMAP(cds, cell_color = "color", scale_color = c("#377EB8",
  "#4DAF4A", "#E41A1C"))
}
\arguments{
\item{cds}{a cell data set}

\item{cell_color}{the parameter in pData you'd like to color your cells by}

\item{scale_color}{a vector of colors you'd like to use for your data set}

\item{nComp}{a vector consisting of the number of PCs to be used}
}
\value{
Returns a UMAP visualization colored by "cell_color" using the colors indicated in "scale_color"
}
\description{
Generate a UMAP visualization from read depth normalized expression matrix
}
\examples{

}
