% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_3d_plot.R
\name{rotate_3d_plot}
\alias{rotate_3d_plot}
\title{Returns a rotating 3d UMAP}
\usage{
rotate_3d_plot(cds = NULL, color_cells_by = NULL)
}
\arguments{
\item{cds}{a cell data set object}

\item{color_cells_by}{a character vector from pData to group cells by}
}
\value{
Returns a rotating 3d UMAP
}
\description{
Returns a rotating 3d UMAP
}
