% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scnorm.R
\name{scnorm}
\alias{scnorm}
\title{Normalize gene counts by sequencing depth per cell}
\usage{
scnorm(cds)
}
\arguments{
\item{cds}{a cell data set}
}
\value{
Returns an expression matrix that is the log of the normalized by sequencing depth converted to counts per 10K with a pseudocount of 1
}
\description{
Normalize gene counts by sequencing depth per cell
}
\examples{

}
