% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myUMAPPlotMarkers.R
\name{myUMAPPlotMarkers}
\alias{myUMAPPlotMarkers}
\title{Given an exisiting UMAP embedding of scRNAseq data, returns UMAP embedding colored by the relative expression of marker gene of interest}
\usage{
myUMAPPlotMarkers(cds, markers = NULL, logMode = T, shape_by = NULL,
  scaled = FALSE)
}
\arguments{
\item{cds}{a cell data set}

\item{markers}{a list of gene short names}

\item{logMode}{whether the log of expression data will be used}
}
\value{
Returns a facet wrapped UMAP embeddings of the same number as the number of marker genes chosen, colored by relative gene expression
}
\description{
Given an exisiting UMAP embedding of scRNAseq data, returns UMAP embedding colored by the relative expression of marker gene of interest
}
\examples{

}
